Update index.html
The current setup uses the browser's short-term memory. I'm going to modify this so that the application uses the browser's long-term memory (localStorage).

Any changes that are made - adding a fix, deleting a comment, changing a status - will now be immediately saved to the LocalStorage of the browser. When you reload the page, the most recent data is automatically reloaded.